move label and type out of type_constraint
