+++ rust/pattern/constraint/type_constraint/mod.rs
@@ -20,9 +20,6 @@
  *
  */
 
-use crate::{TypeVariable, TypeVariableBuilder, UnboundVariable};
-use std::fmt;
-
 pub mod label_constraint;
 pub use label_constraint::*;
 
@@ -40,57 +37,3 @@ pub use relates_constraint::*;
 
 pub mod sub_constraint;
 pub use sub_constraint::*;
-
-#[derive(Debug)]
-pub enum Type {
-    Label(Label),
-    Variable(TypeVariable),
-}
-
-impl Type {
-    pub fn into_variable(self) -> TypeVariable {
-        match self {
-            Self::Label(label) => UnboundVariable::hidden().type_(label).unwrap(),
-            Self::Variable(var) => var,
-        }
-    }
-}
-
-#[derive(Debug, Clone, Eq, PartialEq)]
-pub struct Label {
-    scope: Option<String>,
-    name: String,
-}
-
-impl From<&str> for Label {
-    fn from(name: &str) -> Self {
-        Label::from(String::from(name))
-    }
-}
-
-impl From<String> for Label {
-    fn from(name: String) -> Self {
-        Label { scope: None, name }
-    }
-}
-
-impl From<(&str, &str)> for Label {
-    fn from((scope, name): (&str, &str)) -> Self {
-        Label::from((scope.to_string(), name.to_string()))
-    }
-}
-
-impl From<(String, String)> for Label {
-    fn from((scope, name): (String, String)) -> Self {
-        Label { scope: Some(scope), name }
-    }
-}
-
-impl fmt::Display for Label {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        if let Some(scope) = &self.scope {
-            write!(f, "{}:", scope)?;
-        }
-        write!(f, "{}", self.name)
-    }
-}
